+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- The interdiff of an updated commit is now collapsed behind a spoiler, so that a long interdiff doesn't take up the entire page (https://github.com/svenvanheugten/gh-post-range-diff/pull/71)
+
 ## [0.5.1] - 2026-09-07
 
 ### Added